refactor : Enum class 용도에 따른 분리
